+++ src/app/core/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { appDefaultItemRoute, itemStringUrl } from '../shared/helpers/item-route';
+import { appDefaultItemRoute, itemStringUrl } from '../shared/routing/item-route';
 import { PageNotFoundComponent } from './pages/page-not-found/page-not-found.component';
 
 const routes: Routes = [
