+++ src/app/core/common/item-nav-menu-data.ts
@@ -1,4 +1,4 @@
-import { ItemRoute } from 'src/app/shared/helpers/item-route';
+import { ItemRoute } from 'src/app/shared/routing/item-route';
 import { ItemDetails } from 'src/app/shared/services/current-content.service';
 import { NavMenuItem } from '../http-services/item-navigation.service';
