+++ src/app/shared/routing/content-route.ts
